clean up, move files
