+++ src/coauthor_interface/thought_toolkit/level_3_plugins.py
@@ -1,4 +1,4 @@
-from coauthor_interface.backend.PluginInterface import (
+from coauthor_interface.thought_toolkit.PluginInterface import (
     Intervention,
     InterventionEnum,
     Plugin,
